add percy badge to readme.md; add landing page selenium test
