+++ package.json
@@ -23,5 +23,9 @@
   "dependencies": {
     "babel-cli": "^6.26.0",
     "babel-preset-react-app": "^3.1.2"
+  },
+  "devDependencies": {
+    "@percy/agent": "^0.26.2",
+    "@percy/script": "^1.0.3"
   }
 }
